requirements.txt
# MCP Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# LLM
google-generativeai>=0.3.0

# Database & Storage
aiosqlite==0.19.0

# Logging
structlog==24.1.0

# CLI
click==8.1.7
rich==13.7.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.26.0

# Google Services (optional)
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
